+++ xknx/remote_value/remote_value_control.py
@@ -52,7 +52,6 @@ class RemoteValueControl(RemoteValue[DPTBinary, Any]):
 
     def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTBinary | None:
         """Test if telegram payload may be parsed."""
-        # pylint: disable=no-self-use
         return payload if isinstance(payload, DPTBinary) else None
 
     def to_knx(self, value: Any) -> DPTBinary: