+++ xknx/remote_value/remote_value_switch.py
@@ -44,7 +44,6 @@ class RemoteValueSwitch(RemoteValue[DPTBinary, bool]):
 
     def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTBinary | None:
         """Test if telegram payload may be parsed."""
-        # pylint: disable=no-self-use
         return payload if isinstance(payload, DPTBinary) else None
 
     def to_knx(self, value: bool) -> DPTBinary: